Added the -c|--clear-all-output option to ./replay.sh to clear all output files.
Before this commit, when calling ./sauron.py on a single file, it was
overwriting its output.txt file even if it was specified not to do so.
--> Now the problem is solved.

In this version of the project, every single fit is stored in a single
output.txt file and ./replay.sh add everything in the general
output.txt.
In this way one can handle every single peak separately.

+++ replay.sh
@@ -5,6 +5,8 @@
 # Use -j N to specify the number of parallel process to use 
 #
 
+REWRITE_OUTPUT=0
+
 while :; do
   case "$1" in
     -r|--redo-all-fit)
@@ -27,20 +29,39 @@ while :; do
           ;;
       esac
 
+      REWRITE_OUTPUT=1
+
       break
       ;;
 
-    -s|--single-spectra)
+    -sf|--single-folder)
 
-    #  if [ -d spectra/"$2" ]; then
+      if [ -d spectra/"$2" ]; then
         ./sauron.py $2
-        break
-        ;;
-    #  else
-    #    echo "${2} is not a valid directory"
-    #    exit
-    #    ;;
-    #  fi
+        REWRITE_OUTPUT=1
+      else
+        echo "${2} is not a valid directory --> Exit"
+        break 
+        exit
+      fi
+
+      break
+      ;;
+
+    -c|--clear-all-output)
+      echo "Clearing all output.txt file"
+      for DIR in $(ls spectra); do
+        if [ -d spectra/${DIR} ]; then
+          for file in $(ls spectra/${DIR} | grep ".out.txt");do
+            echo "deleting file ${DIR}/${file}"
+            rm ~/Desktop/Mordor/spectra/${DIR}/${file}
+          done
+        fi
+      done
+
+      break
+      exit
+      ;;
 
     -h|--help)
       echo "Usage: ./replay.sh [OPTION]"
@@ -48,7 +69,7 @@ while :; do
       echo "Example 2: ./replay.sh -s 1157.0208"
       echo "\n"
       echo "-h | --help                 print this help message"
-      echo "-s|--single-spectra         replay.sh will recalculate the fit calling sauron.py on all the spectra the specified sub-folder"
+      echo "-sf|--single-folder         replay.sh will recalculate the fit calling sauron.py on all the spectra of the specified sub-folder"
       echo "-r | --redo-all-fit         replay.sh will recalculate all the fit calling sauron.py (default -j 1)"
       echo "-j N                        if -r is specified use -j N to run sauron.py on N parallel processes"
       echo "\n"
@@ -63,18 +84,21 @@ while :; do
   shift
 done
 
-if [ -f output.txt ]; then
-  echo "--> output.txt already existing\n --> deleting old one\n"
-  rm output.txt
+if [ $REWRITE_OUTPUT -eq 1 ]; then
+  if [ -f output.txt ]; then
+    echo "--> output.txt already existing\n --> deleting old one\n"
+    rm output.txt
+  fi
+  
+  for DIR in $(ls spectra); do 
+    if [ -d spectra/${DIR} ];then 
+      for file in $(ls spectra/$DIR | grep ".out.txt"); do
+        echo "writing ${DIR}/${file} to output.txt"
+        cat ~/Desktop/Mordor/spectra/${DIR}/${file} | tee -a ./output.txt >> /dev/null
+      done
+    fi
+  done;
 fi
-
-for DIR in $(ls spectra); do 
-if [ -d spectra/${DIR} ]
-then 
-  echo "writing ${DIR} to output.txt"
-  cat ~/Desktop/Mordor/spectra/${DIR}/${DIR}.out.txt | tee -a ./output.txt >> /dev/null
-fi
-done;
 
 sed -e '2,${/Integral/d}' -i output.txt
